+++ src/Client/Configuration.hs
@@ -20,11 +20,13 @@ module Client.Configuration
   , ConfigurationFailure(..)
   , configDefaults
   , configServers
+  , configNickPalette
 
   -- * Loading configuration
   , loadConfiguration
   ) where
 
+import           Client.IdentifierColors
 import           Client.Configuration.Colors
 import           Client.ServerSettings
 import           Control.Applicative
@@ -39,6 +41,8 @@ import           Data.Text (Text)
 import qualified Data.Text as Text
 import qualified Data.Text.IO as Text
 import           Data.Traversable
+import           Data.Vector (Vector)
+import           Graphics.Vty.Attributes (Color)
 import           Irc.Identifier (Identifier, mkId)
 import           Network.Socket (HostName)
 import           System.Directory
@@ -49,8 +53,9 @@ import           System.IO.Error
 -- server configuration from '_configServers' is used where possible,
 -- otherwise '_configDefaults' is used.
 data Configuration = Configuration
-  { _configDefaults :: ServerSettings -- ^ Default connection settings
-  , _configServers  :: HashMap HostName ServerSettings -- ^ Host-specific settings
+  { _configDefaults :: !ServerSettings -- ^ Default connection settings
+  , _configServers  :: !(HashMap HostName ServerSettings) -- ^ Host-specific settings
+  , _configNickPalette :: !(Vector Color)
   }
   deriving Show
 
@@ -144,6 +149,9 @@ parseConfiguration def = parseSections $
      _configServers  <- fromMaybe HashMap.empty
                     <$> sectionOptWith (parseServers _configDefaults) "servers"
 
+     _configNickPalette <- fromMaybe defaultNickColorPalette
+                    <$> sectionOptWith parseColors "nick-color-palette"
+
      return Configuration{..}
 
 parseServers :: ServerSettings -> Value -> ConfigParser (HashMap HostName ServerSettings)
@@ -188,8 +196,6 @@ parseServerSettings !def =
        _ssFloodPenalty   <- fieldReq ssFloodPenalty   "flood-penalty"
        _ssFloodThreshold <- fieldReq ssFloodThreshold "flood-threshold"
        _ssMessageHooks   <- fieldReq ssMessageHooks   "message-hooks"
-       _ssNickColorPalette <- fieldReq' ssNickColorPalette
-                                (sectionOptWith parseColors "nick-color-palette")
        return ServerSettings{..}
   where
     field    l key = field'    l (sectionOpt key)