+++ src/Client/State.hs
@@ -68,7 +68,6 @@ import           Client.ChannelState
 import           Client.Configuration
 import           Client.ConnectionState
 import qualified Client.EditBox as Edit
-import           Client.IdentifierColors
 import           Client.Image.Message
 import           Client.Message
 import           Client.NetworkConnection
@@ -234,8 +233,8 @@ recordChannelMessage network channel msg st =
       , rendPalette    = palette
       }
 
-    palette = fromMaybe defaultNickColorPalette
-            $ preview (clientConnection network . csSettings . ssNickColorPalette) st
+    palette = view (clientConfig . configNickPalette) st
+
     -- on failure returns mempty/""
     possibleStatusModes = view (clientConnection network . csStatusMsg) st
     (statusModes, channel') = splitStatusMsgModes possibleStatusModes channel